requirements.txt
django==1.9.9
celery<4
python-dateutil<3
redis<3